requirements.txt
Django==3.2.12
django-treebeard==4.5.1
djangorestframework==3.13.1
drf-yasg==1.20.0
openapi-codec==1.3.2
pillow==9.0.0
pycodestyle==2.8.0
simplejson==3.17.6
toml==0.10.2
tzdata==2021.5
